refactor: shard service
